+++ src/output/mod.rs
@@ -1,6 +1,6 @@
 mod command_output;
 
-pub(crate) use command_output::{CommandOutput, HELP_TEXT};
+pub(crate) use command_output::{CommandOutput, HELP_TEXT, PubSubAck};
 
 #[cfg(test)]
 mod tests;